+++ types/index.test-d.ts
@@ -113,6 +113,8 @@ expectType<Generator<string, undefined, undefined>>(iterate<string>(stringTree))
 expectType<Generator<string, undefined, undefined>>(iterate<string>(stringTree, 0));
 expectType<Generator<string, undefined, undefined>>(iterate<string>(stringTree, 0, 10));
 expectType<Generator<string, undefined, undefined>>(reverseIterate<string>(stringTree));
+expectType<Generator<string, undefined, undefined>>(reverseIterate<string>(stringTree, 0));
+expectType<Generator<string, undefined, undefined>>(reverseIterate<string>(stringTree, 0, 10));
 expectType<boolean>(equals<string>(stringTree, stringTree));
 expectType<boolean>(equals<string>(stringTree, stringTree, areStringsEqual));
 expectType<boolean>(equals<number, string>(numberTree, stringTree, areNumberAndStringEqual));