+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## v0.16.0 (unreleased)
 
-  * `start` and `end` (index) parameters have been added to `iterate`.
+  * `start` and `end` (index) parameters have been added to `iterate`
+    and `reverseIterate`.
 
 ## v0.15.0 (2025-07-13)